+++ cockatrice/src/server/chat_view/chat_view.h
@@ -2,7 +2,7 @@
 #define CHATVIEW_H
 
 #include "../../client/tabs/tab_supervisor.h"
-#include "../user/user_list.h"
+#include "../user/user_list_widget.h"
 #include "room_message_type.h"
 #include "user_level.h"
 #include "user_list_proxy.h"
