+++ cockatrice/src/server/user/user_list_widget.h
